requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
APScheduler==3.10.4
cachelib==0.13.0
faiss-cpu==1.8.0.post1
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Session==0.8.0
google-cloud-speech==2.27.0
google-cloud-storage==2.18.2
google-generativeai==0.8.1
gunicorn==23.0.0
httpx==0.27.0
Jinja2==3.1.2
langchain==0.2.16
langchain-text-splitters==0.2.4
mem0ai==0.1.15
numpy==1.26.4
openai==1.40.6
pillow==10.4.0
PyQRCode==1.2.1
pytest==7.4.3
python-docx==1.1.2
python-dotenv==1.0.1
qdrant-client==1.11.2
qrcode==7.4.2
requests==2.31.0
sentence-transformers==3.0.1
tiktoken==0.6.0
torch==2.2.1+cpu
torchvision==0.17.1+cpu
torchaudio==2.2.1+cpu
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
redis==5.1.0
google-api-python-client==2.142.0
beautifulsoup4==4.12.3
chardet==5.2.0
# newspaper3k==0.2.8
# lxml-html-clean==0.3.1
# goose3==3.1.19

peewee==3.17.8